fix: utilisation de Long comme type de retour pour countAllBooks et countCategories

+++ src/main/java/sn/iage/isi/repositories/BookRepository.java
@@ -141,10 +141,9 @@ public class BookRepository {
                 .getResultList();
     }
 
-    public int countAllBooks() {
+    public Long countAllBooks() {
         return em
                 .createQuery("SELECT COUNT(b) FROM Book b", Long.class)
-                .getSingleResult()
-                .intValue();
+                .getSingleResult();
     }
 }
